agent/pyproject.toml
[project]
name = "afa-agent"
version = "0.1.0"
description = "afa agent"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "google-adk>=1.16.0,<2.0.0",
    "a2a-sdk[all]~=0.3.9",
    "nest-asyncio>=1.6.0,<2.0.0",
    "opentelemetry-instrumentation-google-genai>=0.1.0,<1.0.0",
    "gcsfs>=2024.11.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "google-cloud-aiplatform[evaluation]>=1.118.0,<2.0.0",
    "fastapi~=0.115.8",
    "uvicorn~=0.34.0"
]



#dependencies = [
#    "google-adk>=1.18.0",
#    "fastapi>=0.115.0",
#    "python-dotenv>=1.0.0",
#    "uvicorn[standard]>=0.32.0",
#    "google-cloud-aiplatform>=1.38.1",
#    "google-cloud-modelarmor",
#    "pandas",
#    "dotenv",
#    "google-adk[a2a]"
#]

#
#[project.optional-dependencies]
#dev = [
#    "pytest>=8.0.0",
#    "pytest-asyncio>=0.24.0",
#]
#
#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"
#
#[tool.uv]
#dev-dependencies = []
#
#[tool.hatch.build.targets.wheel]
#packages = ["app"]
#
#[tool.hatch.build.targets.wheel.shared-data]
#"app/static" = "share/bidi-demo/static"
#"app/.env" = "share/bidi-demo/.env.template"
